feat: create jwt types.ts
