+++ packages/better-auth/src/plugins/jwt/sign.ts
@@ -8,9 +8,10 @@ import {
 import type { GenericEndpointContext } from "../../types";
 import { BetterAuthError } from "../../error";
 import { symmetricDecrypt, symmetricEncrypt } from "../../crypto";
-import { getJwtPlugin, type JwtPluginOptions } from ".";
+import { getJwtPlugin } from ".";
 import type { Jwk } from "./schema";
 import { getJwksAdapter } from "./adapter";
+import type { JwtPluginOptions } from "./types";
 
 /**
  * Signs a payload in jwt format
@@ -79,29 +80,6 @@ export async function signJwt(
 		ctx.context.session?.user.id;
 	if (sub) jwt.setSubject(sub);
 	return await jwt.sign(privateKey);
-}
-
-/**
- * Backwards compatable version of signJwt
- *
- * @deprecated - prefer signJwt
- *
- * @param ctx - endpoint context
- * @param options - Jwt signing options. If not provided, uses the jwtPlugin options
- */
-export async function getJwtToken(
-	ctx: GenericEndpointContext,
-	options?: JwtPluginOptions,
-) {
-	if (!options) {
-		options = getJwtPlugin(ctx.context).options;
-	}
-
-	const payload = !options?.jwt?.definePayload
-		? ctx.context.session!.user
-		: await options?.jwt.definePayload(ctx.context.session!);
-
-	return await signJwt(ctx, payload, options);
 }
 
 export async function generateExportedKeyPair(options?: JwtPluginOptions) {
